+++ Src/Library/Binder/RequestBinder.cs
@@ -400,5 +400,4 @@ public class RequestBinder<TRequest> : IRequestBinder<TRequest> where TRequest :
         }
         return dictionary;
     }
-
 }