+++ Src/Library/Interfaces/IPostProcessor.cs
@@ -10,5 +10,5 @@ namespace FastEndpoints;
 /// <typeparam name="TResponse">the type of the response dto</typeparam>
 public interface IPostProcessor<TRequest, TResponse>
 {
-    Task PostProcessAsync(TRequest req, TResponse res, HttpContext ctx, IReadOnlyCollection<ValidationFailure> failures, CancellationToken ct);
+    Task PostProcessAsync(TRequest req, TResponse? res, HttpContext ctx, IReadOnlyCollection<ValidationFailure> failures, CancellationToken ct);
 }