+++ Src/Library/changelog.md
@@ -2,6 +2,7 @@
 - signature of global error response builder func has changed to include the `HttpContext` #220 #230
 
 ### NEW
+- support .net 7.0 via multi targetting
 - endpoint configuration with groups and sub/nested groups #214
 - `[Throttle(...)]` attribute for configuring endpoints #227
 - `HttpContext.MarkResponseStart()` and `HttpContext.ResponseStarted()` extension methods #230
@@ -27,7 +28,7 @@
 - nre when publishing an event and no handlers are registered #259
 
 ### IMPROVEMENTS
-- upgrade solution/build pipeline to .net 7.0
 - optimize default request binder by reducing allocations.
 - swagger schema resolving
 - json object array string binding of requests from swagger ui
+- remove `notnull` constraint from `TResponse` generic argument of endpoint class